+++ tests/Rule/LatteTemplatesRule/PresenterWithModule/Fixtures/Modules/Foo/Presenters/FooBarPresenter.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Efabrica\PHPStanLatte\Tests\Rule\LatteTemplatesRule\PresenterWithModule\Fixtures\Modules\Foo\Presenters;
+
+use Nette\Application\UI\Presenter;
+
+final class FooBarPresenter extends Presenter
+{
+    public function renderDefault(): void
+    {
+        //
+    }
+}